# Optimizing AttributedSpans.collapseSpans() to run in linear time