The current implementation of `collapseSpans` operates by pre-generating the list of start and end points for the collapsed spans, then walking that list and repeatedly checking the list of `SpanMarker`s to determine which `Attribution`s should apply to each span. In cases with many spans this causes a large amount of duplicate work, especially in cases with overlapping spans. In my testing it takes the method almost 2 seconds to return with 1000 spans. For a less contrived example, it only needs about 50 overlapping spans to take over 16ms (time for a frame when running at 60fps).

This PR:

- Changes `collapseSpans` to run with a single linear pass of the `SpanMarker` list. It also eliminates the preprocessing step of generating the start and end locations for the collapsed spans, which wasn't significantly contributing to the runtime but fewer LoC is always a nice bonus.
- Updates the sorting logic for `SpanMarker` to always have `start` markers precede `end` markers. Without this the inner loop of `collapseSpans` needs some ugly conditionals to handle an edge case where an end marker precedes a start marker with the same offset.
- Adds a couple tests to the AttributedSpans test suite for behavior that that wasn't already covered but broke tests from other parts of the library while I was working on this.
- Adds some minor additions to `MultiAttributionSpan`: `copyWith` for convenient copying, `toString` override for better log messages.

After these changes it the method takes 3ms to return with 50 spans and 30ms with 1000. I ran the tests locally via `flutter test` and everything passed.